coproc/tests: Have supervisor move batches, not create copies
- There were discrepencies in tests related to use of the record_batch_builder,
when testing with the identity_coprocessor. Differences were tracked down to use
of the record_batch_builder, and upon further inspection, its noted that its not
even necessary to use this here.

- Furthermore, its not necessary to copy records on behalf of the apply()
functor at all. Change the signature of apply() to take a const& vector of
records and the functor can perform the copy if it wishes.