coproc/tests: Modify startup procedure in router_test_fixture
- Makes startup more automatic, since the superclass has the list of
coprocessors that are to be enabled, this can be used to call
register_coprocessors on behalf of caller.

- This change also makes this closer to what would happen in production.
Previously the router would be manually manupulated to hold the registered
copro. Now this class will use a client to make a request to the
coproc::service, just like how things would work in prod.