coproc: pacemaker_context place for shared state across scripts
- The pacemaker context is the first step in the new pacemaker architecture. It
is a place where state and methods that are to be shared across a soon to be
implemented script model (per shared) can be.

- There will be one pacemaker_context per shard, and a reference to it will be
passed to each script_context that spawns on that same shard. The
pacemaker_context will be destroyed when the coproc service shutsdown, i.e. when
redpanda shuts down.