coproc: scheduled_task - mechanism for deploying coproc fibers
- In the new coproc architecture we plan to move away from a single fiber (aka
run-loop) model to a multi fiber model with the unit of context begin a coproc
script.

- A 'scheduled_task' is the backing mechanism for how the fiber run loop
operates. Consumers of this interface just need to implement 'do_task()' and the
virtual method will be called at the periodicity defined in the constructor. No
need to worry about timers and/or scheduling

- A gate guards the resources and ensures that no member variables will be
accessed outside of the gate, so that proper shutdown is guaranteed.