coproc: script_context - new base unit of the coprocessor framework
- The script_context is the smallest schedulable unit in the coprocessor
framework. It represents one logical fiber. It is split across shards as
needed (for each interested ntp if that ntp exists on said shard, there will be
a script_context on that shard for it)

- The script context manages a single async run loop which runs its actions
synchronously, i.e. the read -> send -> write loop is single threaded and those
actions will occur in that order.

- Higher concurrency is achieved by having multiple script contexts exist at the
same time. They all operate independently of eachother, and it is in this
fashion that we take more advantage of the seastar reactor model. Compared to
the previous implementation where there was one single threaded loop context and
a much lower useage of concurrency. Work that could have been done was wasted
while one script was waiting on a particular action to complete.

- Fixes #318