rptest/wasm: WasmTest class for writing ducktape tests
- Wraps the repetitve part of writing any wasm test, starting a script, pushing
input data and expecting a set amount of output data on a set of defined topics.
Many of these parameters depend on the script that is deployed as well.

- To use, simply pass a test plan (a list of wasm_scripts) and the class will
build your scripts, deploy them, and produce a set amount of data (that you
provide).

- It then waits on an expected amount of data (that you provide) and once read,
forwards all input and output records to the test writers assertion method.