archival: Do not re-upload topic manifest
Previous behavior: topic_manifest is uploade when new ntp is added
to the archival service. This leads to the situation when it gets
uploaded from several nodes at the same time. On GCS this can be
throttled since it can only allow one update per second on the
same object.

New behavior: topic_manifest is uploaded when new ntp is added and
partition id is 0. The partition 0 is guaranteed to exist and we
only upload from the leader. This guaranees that we're only uploading
topic_manifiest from one place in the cluster.