r/consensus: do not decide about committed index update if not a leader
Recent changes in raft causing `log::flush()` to be executed outside the
semaphore caused a problem which might lead to the situation in which
one of the replicas was not able to make progress. The problem was
caused by the fact that `_committed_index` was updated on the node after
it was not longer a leader.

Signed-off-by: Michal Maslanka <michal@vectorized.io>