coproc: Sync materialized topic term w/ input
- Through refactoring tests a fatal bug was observed. If a coprocessor
reorders records, and one with a lower term_id superceeds another with a
higher, coproc will blindly write this to storage, and storage will
assert.

- This change modifies the term id of transformed batches to match the
latest in the input log. This way the latest term of a materialized
topic will always match the input log, and the described crash above can
never occur again.