c/metadata_dissemination: do not send dissemination requests to self
When partition is moved from the node metadata gets updated first after
the update it may happen that current node id not longer part of
partition replica set but it still hosts one of the replicas, this
replica may generate metadata updates. In this case we want to skip
sending request to current node since it already received an updated
(via leadership notification).

Signed-off-by: Michal Maslanka <michal@vectorized.io>