archival: Use dirty offset of the log segment
Archival upload algorithm uses committed offset of the storage layer.
It expects that committed offset of the segment is consistent with base
offset of the next segment. But currently, this isn't always the case
(as described in #1521). This situation leads to the simultaneous upload
of the same segent which triggers throttling (as described in #1800).

This commit switches archival subsystem from committed storage layer
offset to dirty offset. The dirty offset of the storage layer is always
consistent with respect to the base offset of the next segment. The
archival service uses high
watermark raft offset to decide if the segment can be uploaded. It never
uploads a segment that contains an offset which is higher than high
watermark.

This change makes archival service tolerant to a situation when the
committed offset is inconsistent (as described in #1800).

This change might introduce the following problem: when the log is
truncated on a leader the already uploaded log segment might be
truncated as well (since we upload by dirty offset). But check against
high watermark should prevent this situation.