c/members_manager: checking if config update is required after join
When node tries to join the cluster it may happen that it was already
part of it but the node data folder was wiped out. In this scenario node
with removed data will try to join the cluster until it will receive
updates from rest of the nodes. After this update is received node that
is trying to join the cluster will stop joining procedure since it is
already a cluster member. After that joining node must check if
configuration update is require since its configuration might have been
changed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>