raft: add temporary workaround for recovery failure
This is a temporary workaround that adds the configuration parameter

  full_raft_configuration_recovery_pattern

that specifies one or more patterns that are matched against NTPs as
a prefix filter. If any NTP matches then upon starting that NTPs raft
group recovery will replay the entire log rather rather than relying
on the stored configuration snapshot used as a recovery optimization.

The configuration is a string (or array of strings) matches as
prefix to each "namespace.topic.partition" (e.g. `redpanda.controller`,
and if "*" is contained in the set of patterns all ntps will match.

Details about the bug that this workaround addresses:

----

When raft starts it reads a group configuration snapshot from the key-value
store, and then replays the log from the snapshot position. The idea here is to
make start-up faster by not replaying the entire log to find all of the active
configurations.

Consider a configuration stored in the key-value store that is loaded on start:

```
Starting raft bootstrap from 195
...

// The configuration:
{ offset: {-9223372036854775808}, idx: {0},cfg: {current: {voters: {}, learners: {}},
{ offset: {0}, idx: {1},cfg: {current: {voters: {{id: {0}, revision:
{ offset: {1}, idx: {2},cfg: {current: {voters: {{id: {0}, revision:
...
```

The contents probably don't matter so much to the general problem, but in the
case we'll layout the conflict arises with offset 0.

The next thing that happens during startup is that the log is replayed to
generate a raft bootstrap state that contains some stats about how many
configurations were found as well as offsets of interest.

```
trying to get reader for: {start_offset:{196}, max_offset:{196}, min_bytes:0,
reader cache miss for: {start_offset:{196}, max_offset:{196}, min_bytes
adding reader [196,196]
...
Read bootstrap state: data_seen 0 config_seen 1 eol false commit 196 term 0 prev_idx 0 prev_term 0 config_tracker 196 commit_base_tracker 196
```

The important bits here are the `config_seen 1` and `prev_idx 0` which were
computed by replaying the log from the configuration snapshot location forward
to the end of the log.

The next bit of code is where the problem arises: we saw a configuration batch
(config_seen > 0) and the prev_idx is 0. When the configuration is added to the
configuration manager at offset 0 then an exception is thrown, because as we
saw above, a configuration already exists at this offset.

```
if (st.config_batches_seen() > 0) {
_configuration_manager.add(
   st.prev_log_index(), st.release_config());
}
```

So, what could have led us to this point? There may be many cases, but here is
one:

When the bootstrap configuration is being created we scan the log and process
record batches. When a configuration batch is encountered this happens where we
bump the config_seen count, and process_offsets:

```
33     auto last_offset = b.last_offset();
 34     if (_log_config_offset_tracker < last_offset) {
 35         _config_batches_seen++;
 36         _log_config_offset_tracker = last_offset;
 37         process_offsets(b.base_offset(), last_offset);
```

process_offsets in turn initializes the prev_idx to commit_index (which is initially 0).

```
58 void configuration_bootstrap_state::process_offsets(
 59   model::offset base_offset, model::offset last_offset) {
 60     // happy path
 61     if (last_offset > _commit_index) {
 62         _prev_log_index = _commit_index;
 63         _prev_log_term = _term;
 64         _commit_index = last_offset;
 65         _commit_index_base_batch_offset = base_offset;
 66         return;
 67     }
```

This means that if no other batches are processed--that is, the log contains
exactly one batch which is a config batch, then prev_idx would remain as 0 and
config_seen > 0 as we saw above.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>