kafka/server: Modify partition_proxy creation
- Previously to know what impl of partition_proxy to instantiate, a
method would parse the topic from an incoming ntp to determine if a
topic was materialized or not, depending on the now deprecated
materialized_topic naming convention.

- Now that this has been deprecated make_partition_proxy must query
topic metadata to know if a topic is materialized or not.

- Adding an is_leader method to the interface as well. For a
"materialized_partition" it is the leader if its source topic is also a leader.