cloud_storage: Add remote_segment component
The component is supposed to be used by the remote_partition component
that will be added next. It is responsible for downloading segment to
the cache. It can also be used to get a ss::input_stream that returns
segment data. The instances of this class are created lazily and then
destroyed when they're no longer needed.

The remote_segment_batch_reader component is an object that can be used
to retrieve individual batches from the segment.

There is also a remote_segment_batch_consumer component which is used
with storage::continous_parser.

The component is used for offset translation. The offset translation in
shadow indexing is tricky. The data which is used by offset translation
algorithm is sparse. We're only storing an offset delta for base_offset
of every segment. This means that we can't precisely translate arbitrary
offset.

In order to bypass this limitation the remote_segment_batch_consumer
maintains a running offset delta which is updated for every record
batch. It can be used to translate an offset of the current record
batch. The log_reader_config is storing kafka offsets instead of
redpanda offsets. It also extended to have additional field to store
start_offset_redpanda (which is needed by remote_partition).