cluster: Decouple revision check from log remove
- The state of the shard table and how it represents an underlying log
can get out of sync.

- This is due to the fact that the shard table is being updated out of
band on all cores, so it could rarely occur that there would be no entry
for the log to be removed, because another shard just removed the entry.

- The effect would be the underlying log stays, and a subsequent create
call raises an assertion due to the fact that an unexpected log has been
detected.

- The solution is to decouple the removal from the if condition that was
surrounding it. Obtaining the log from the log manager should only work
on the shard that the log exists on.

- Fixes #2726