coproc/tests: Read until exact record count
- Due to the introduction of the kafka::client in the coproc/tests
harness, an unforceen side effect was introduced. Re-batching is
performed by the client and coproc can no longer assert the number of
record batches.

- Sometimes tests were passing when they shouldn't have because the
number of expected record batches was observed, but the number of
records was not.

- The only metrics coproc/tests should be using to assert if a test is
correct is to compare the exact number of records.

- To do this the consume() API was again modified to take a limit
parameter. This is when the consumption loop stops, when the limit
reaches the records_read local var, or if timeout has been reached.

- Call sites have been modified to use a new utility method that makes a
record batch reader with the exact records specified.