coproc/tests: Remove idle wait fiber solution
- The solution made readers wait until all copro fibers were in an idle
state before reading. The intention was to know at what point would a
coprocessor be finished processing, or i.e. be up to date with all of
its input topics.

- This mechanism was introduced to make consume() calls wait on this
action. Once the idle state was detected a `list_offsets` request could
be made to find the "end" of the log.

- However issues such as timeouts can cause fibers to return to an idle
state, making tests think the coprocessor has finished processing when
it does have more data to process.