raft/consensus: quiet, clarify waiting for recovery msg.
As a new user, when I saw a WARN log message about waiting
for a node to recover, I thought that one of the RP nodes
was failing.

To make this less alarming, downgrade to INFO, and use
phrase "catch up" instead of "recover". "recover" is more
consistent with code, but I think this is clearer to your
average user that doesn't know our code.