cluster: topic_table events grouping
- Modifying the topic table to have two sets of events and listeners,
one for non_replicable topics, and one for normal ones

- To reduce code duplication and keep the PR to the minimal number of
changes all existing wait / update logic was unmodified and instead
wrapped in a new templated struct called 'wait_notification'

- One instance of this struct is made for each notification type

- Emitters of events, add to the correct instance of this struct, and
call the `notify_waiters()` event on the correct instance when they are
finished.

- After this change, redpanda should behave exactly as it did before.
The new abstraction allows for future commits to build on this.