coproc: Reconciliation backend
- The reconciliation backend will mimmic the
cluster::controller_backend.

- Its purpose is to handle all non_replicable* op type events that are
emitted from the topics table.

- The rational for this is that it is not possible for coproc to
synchronize around the handling of these op_types if they are handled by
code in cluster. It can only know that the event was emitted, not when
it was finished.

- With this solution the complexity of the existing
cluster::controller_backend is reduced and more coproc code can stay in
the coproc namespace