utils: Protect from compiler bug in clang on ARM
Capture state in lambda function inside 'retry_chain_logger::log' method.
Pass the lambda to the 'do_log' method that performs the actual logging.
This is needed to prevent clang compiler bug from manifesting. The bug
is triggered when the code in the log method is inlined. To prevent
inlining the 'do_log' method has 'noinline' attribute'. The body of the
'log' method can be inlined but it doesn't do anything that can trigger
the bug anymore.

The log method works the same way as seastar::logger::log which is immune
to this compiler bug.