kafka/client: Copy func in retry_with_mitigation
This fixes a bug where lambda captures are not available for during
the retry due to the lambda being moved from. E.g., a captured
topic_partition would have no topic name as the string has been moved.

During a release mode build, `SEASTAR_TYPE_ERASE_MORE` is not defined,
this results in the following code bing executed for a continuation:
```
    then(Func&& func) noexcept {
        return call_then_impl::run(*this, std::move(func));
    }
```

When `SEASTAR_TYPE_ERASE_MORE` is defined, the capture is by value,
although that doesn't seem to be the intention of the author:
```
    ncf = noncopyable_function<func_type>([func = std::forward<Func>(func)](auto&&... args) mutable {
        return futurize_invoke(func, std::forward<decltype(args)>(args)...);
    });
```

The change was made in: https://github.com/scylladb/seastar/commit/0dab3242f4717f892ccdc0595f836c0666b214f8

Without this fix, there are many failures for schema registry and
pandaproxy ducktape tests when the next commit is applied.

Signed-off-by: Ben Pope <ben@vectorized.io>