raft: simplify node removal tests
These tests were polling on the underlying
log behind the consensus instance for the node
that was removed.

This diverges from how the code is used in
real life (the consenseus object for a removed
node is torn down by the controller, not left alive)
and could cause weird hangs in the test when
the orphaned consensus instance was left in
a just-truncated state.

Simplify the tests to just check the visible offset
at the raft layer to validate that the removed
node hasn't advanced, rather than reading its underlying
storage.

Fixes: https://github.com/vectorizedio/redpanda/issues/2799

Signed-off-by: John Spray <jcs@vectorized.io>