js: Pass copy of input to apply() method
- Previously the apply() method was recieving a record_batch that upon a
call to a function called map() returned new records with cleared out
fields.

- This change ensures we pass a deep copied recordBatch to each
invocation of apply. Deep copies are necessary due to the fact that
multiple Coprocessors may be the reciepient of this payload, and its not
desired to  mutations of value fields persist across invocations of
apply()

- Also this change will now allow clients to observe all of the values
of the properties that exist in recordBatch and its nested types.
Previously upon query zeroed out values were observed.