kafka/p: Introduce flex support for our generator
- Our kafka codegenerator explicity left out flex support

- The strategy taken to accomplish the task was to have the code
generator, generate read_flex* and write_flex* calls instead of the
respective read* and write* calls.

- A small optimization is included where the generated code will check
once (at entry of encode/decode) weather the request should be interpreted
as a flex request. If true then a path will be taken to another method
which uses the proper flex versions of all read* & write* calls. If
false the previously generated stubs will be used.