kafka/generator: Remove un-used conditional case
- This commit ensures that when arrays of integral types are marked as
'nullable', that their contents are not also interpreted as 'optional'.

- This condition can never occur anyway. There is no way within the
generator syntax to create an array of an integral type, where the
fields are optional.

- To test this case occuring you can add a 'nullableVersions' key
to a type like `[]int32` an example would be in
`add_partitions_to_txn_request.json`  Setting the nullableVersions
field makes the array optional, but without this commit, also makes
the fields within the array optional as well.

- Before this commit (does not compile):
```
        v.partitions = reader.read_nullable_array([version](request_reader& reader) {
            {
                auto tmp = reader.read_int32();
                if (tmp) {
                    return model::partition_id(std::move(*tmp));
                }
                return std::nullopt;
            }
        });
```

- After:
```
        v.partitions = reader.read_nullable_array([version](request_reader& reader) {
            return model::partition_id(reader.read_int32());
        });

```