rptest: add failure injection Python types
This commit adds types that are meant to be used with Redpanda's low
level failure injection functionality (see
`storage_failure_injection_config_path`). The intention is for tests to
define their failure injection requirements using these types and to
serialise the config to a JSON file which will be picked up by Redpanda.