rptest: add a redpanda monitoring service
Normally, Redpanda nodes are not expected to stop unexpectedly when
running in ducktape. When running tests with failure injection enabled,
this is no longer true: if an EIO is injected, Redpanda should assert
out. In order to support such tests, this commit introduces a
straight-forward ducktape monitoring service. It polls all Redpanda
nodes that are supposed to be running periodically and restarts the ones
that have crashed.