Add unhealthy reason to ClusterHealthOverviewTest
Adds unhealthy reasons checking to the cluster health test.

Basically we check the invariant that this list is empty when the
cluster is health, and that it contains the expected reasons for
the unhealthy operations covered by the tests, which include node
down and under-replicated partitions.

We check that the expected reason is only one of the reasons, not the
only reason, since other reasons are likely to appear, perhaps
transiently: e.g., after a node goes down, we will expect to have some
under replicated partitions, and perhaps, briefly, some leaderless
partitions.

Issue #5263.