archival: block upload loop mid-upgrade
We disable the upload loop while in a mixed version state. If we have to
rollback, we can end up with a system where the archival stm is stuck on
all nodes (the only way to get unstuck would be to finish the upgrade).
The system would look fine: can produce and consume, but uploads
wouldn't proceed and it leaves us vulnerable to disks filling up.