acl: Avoid allocations when checking implied ops
`acl_implied_ops` was quite visible (~3%) in the callchains leading to
`operator new()` in the LRC profile.

It's called for every partition so this is not entirely surprising when
using larger partition counts.

Avoid the allocations by "inlining" the check into the function so that
no vector has to be allocated.