transactions: fix ListTransactions return open transactions
Clients use scatter gather approach and asks all brokers to list
transacitons. A broker is expected to check hosted txn coordinators
only if the broker is the current leader of the coordinators and let
a client to merge the results. Fixing list transactions handler to
ignore not leader errors.

Fixes https://github.com/redpanda-data/redpanda/issues/11948