cloud_storage: archive retention from clean offset
Previously, sized based retention within the archive could delete extra
segments under certain scenarios. This is due to the fact that the
archive size is updated along with the clean offset instead of the start
offset.

Consider the following case:
    C       S
|_|_|_|_| |_|_|_|_| |_|_|
^         ^         ^
Spill 1 Spill 2    STM

In the example below, we have 2 spillover manifests with 4 segments
each. The archive clean offset (C), is at the start of the third segment
in the first spill and the archive start offset (S) is at the start of
the second segment in the second spill manifest.

Assume all segments have the same size. Let's say retention is 4
segments (retention = 4). At this point the cloud log size is of 8
segments (cloud_log_size=8). Retention thinks 4 segments should be
removed (to_remove=cloud_log_size - retention=4). With the code in dev,
we start counting from the first segment in spillover 2. The truncation
point we get is the start of the STM manifest. This is incorrect as we
only end up keeping 2 segments

The fix here is to start counting retention eligible segments from the
clean offset. In order to achive this, I had to tweak the cursor
fetching logic to allow operating below the archive start offset.