rpk cloud byoc: pass --redpanda-id to the plugin
This previously was implicitly always passed because rpk did not strip
rpk specific flags.

A PR in April changed the behavior to strip all rpk specific flags --
this was done for safety so that people could pass -X, -v, etc. to rpk
while also execing the plugin, and rpk would strip all rpk flags and
only give the plugin the plugin specific flags.

Since --redpanda-id was a defined rpk flag, it wasn't obvious that this
flag was a part of the contract handoff to byoc, and we did not include
it in the PR.

As it turns out, the flag is required.