c/partition_balancer: force-refresh health monitor only once
If partition sizes are missing, we force-refresh the health monitor. The
problem is that this doesn't guarantee that they will appear (e.g. if
all partition replicas are on dead nodes, they won't). In this case the
balancer will try to plan in a busy-loop. To fix this, do force-refresh
only once and if it doesn't help, revert to sleeping before the next
tick.

Fixes https://github.com/redpanda-data/redpanda/issues/12307