log_eviction_stm: further simplifications
Removes the queue-based signaling in log_eviction_stm, which proved to
be a bit error prone.

Instead, this reverts to the old implementation using a
condition_variable, but leaves out the offset monitor, which doesn't
seem needed.

Now, either storage or delete-records will bump their respective
offsets, and the event-processing loop will take the higher and truncate
to it. If this isn't finished, the process is repeated after some
timeout.