ducktape/cloud_storage: Ignore trim failure in test with small cache
In certain tests when the cache size is very small and trimming is done
with a lot of downloads in progress, it is possible that cache trimming
will not be able to free up enough space. This is usually a problem for
systems with a well configured cache, but for tests with a small cache
these errors should be ignored.